Move go files into actions/helpers packages
