+++ actions/index.go
@@ -1,0 +1,48 @@
+package actions
+
+import (
+	"context"
+	"log"
+	"net/http"
+
+	h "github.com/jmcarp/deploy-to-cf/helpers"
+
+	"github.com/google/go-github/github"
+	"github.com/gorilla/csrf"
+	"github.com/gorilla/schema"
+	"golang.org/x/oauth2"
+)
+
+func Index(c *h.Context, w http.ResponseWriter, r *http.Request) {
+	source := Source{}
+	if err := schema.NewDecoder().Decode(&source, r.URL.Query()); err != nil {
+		log.Println(err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	client := github.NewClient(nil)
+	app, err := h.LoadManifest(client, source.Owner, source.Repo, source.Ref)
+	if err != nil {
+		log.Println(app, err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	session, _ := c.Store.Get(r, "session")
+	token := session.Values["token"].(oauth2.Token)
+	authClient := c.OauthConfig.Client(context.TODO(), &token)
+	targets, err := h.FetchTargets(authClient, c.Config)
+	if err != nil {
+		log.Println(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	c.Templates.ExecuteTemplate(w, "index.html", map[string]interface{}{
+		csrf.TemplateTag: csrf.TemplateField(r),
+		"App":            app,
+		"Source":         source,
+		"Targets":        targets,
+	})
+}
