+++ helpers/archive.go
@@ -1,4 +1,4 @@
-package main
+package helpers
 
 import (
 	"archive/tar"
